refactor: 重构 electron 入口文件
